:sparkles: [FEAT] 사용자의 여행 등급 조회 기능 구현
