+++ src/main/java/com/outofcity/server/global/exception/message/SuccessMessage.java
@@ -18,7 +18,8 @@ public enum SuccessMessage {
     ACTIVITY_LIKE_ADD_SUCCESS(HttpStatus.CREATED.value(), "액티비티 찜에 성공하였습니다."),
     ACTIVITY_LIKE_DELETE_SUCCESS(HttpStatus.OK.value(), "액티비티 찜 취소에 성공하였습니다."),
     RESERVE_ACTIVITY_READ_SUCCESS(HttpStatus.OK.value(), "예약 액티비티 목록 조회에 성공하였습니다."),
-    COMPLETED_ACTIVITY_READ_SUCCESS(HttpStatus.OK.value(), "체험 완료한 액티비티 목록 조회에 성공하였습니다."),;
+    COMPLETED_ACTIVITY_READ_SUCCESS(HttpStatus.OK.value(), "체험 완료한 액티비티 목록 조회에 성공하였습니다."),
+    RANK_GET_SUCCESS(HttpStatus.OK.value(), "사용자 등급 조회에 성공하였습니다."),;
     private final int code;
     private final String message;
 }
